move component to addon folder
